% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpeaks.R
\name{findpeaks}
\alias{findpeaks}
\title{Find Peaks}
\usage{
findpeaks(
  x,
  nups = 1,
  ndowns = nups,
  zero = 0,
  minpeakheight = -Inf,
  minpeakdistance = 1,
  threshold = 0,
  npeaks = 0,
  sortstr = FALSE
)
}
\arguments{
\item{x}{\code{numeric(n)}. Numerical vector taken as a time series (no \code{NA}s allowed).}

\item{nups}{\code{integer(1)}. Minimum number of increasing steps before a peak is reached.}

\item{ndowns}{\code{integer(1)}. Minimum number of decreasing steps after the peak.}

\item{zero}{\code{integer(1)}. Can be \code{1}, \code{-1}, or \code{0} (the default); how to interpret succeeding steps of the same
value: increasing, decreasing, or special.}

\item{minpeakheight}{\code{integer(1)}. The minimum (absolute) height a peak has to have to be recognized as such.}

\item{minpeakdistance}{\code{integer(1)}. The minimum distance (in indices) peaks have to have to be counted.}

\item{threshold}{\code{integer(1)}. The minimum peak size.}

\item{npeaks}{\code{integer(1)}. The number of peaks to return.}

\item{sortstr}{\code{logical(1)}. Should the peaks be returned sorted in decreasing order of their maximum value?}
}
\description{
Find peaks (maxima) in a time series.
}
\details{
This is a reimplementation of \code{pracma::findpeaks()} which uses a character string and a regex to identify the peaks.
This method is very slow and so this reimplementation using numeric values and the \code{\link[=rle]{rle()}} instead.
}
